requirements.txt
djangorestframework
psycopg2-binary
redis
django-silk